feature: add attempt submitting

+++ mesh-exam/feature/profile/src/main/java/io/flaterlab/meshexam/presentation/profile/dvo/HistoryDvo.kt
@@ -3,6 +3,6 @@ package io.flaterlab.meshexam.presentation.profile.dvo
 internal data class HistoryDvo(
     val id: String,
     val name: String,
-    val durationInMin: Long,
+    val durationInMin: Int,
     val isHosted: Boolean
 )
